+++ demos/aggregation-module-app/src/main/resources/META-INF/properties.xml
@@ -53,7 +53,7 @@
   </property>
   <property>
     <name>dt.application.AggregationModuleRegressionApp.operator.Validator.timeBuckets</name>
-    <value>1m</value>
+    <value>1m,1h,1d</value>
   </property>
   <property>
     <name>dt.application.AggregationModuleRegressionApp.operator.Validator.activeAggregationInterval</name>
